Refactor WelcomeView environment initialization
